requirements.txt
fastapi
uvicorn
python-dotenv
openai>=1.0.0
google-generativeai
llama-cpp-python==0.3.7
serpapi
Pillow
pdfminer.six
python-docx
pymupdf
pdf2image
unstructured
python-docx
pytesseract
python-multipart
mammoth
pybind11>=2.11.1
cmake>=3.15
html2text
google-genai
PyMuPDF
google-search-results
python-json-logger==2.0.7
pydantic-settings>=2.0.0